Very minor restyling with merge.

* origin/pr/16889:
  msvc optimization flags tune
  astyle code format for all files I changed
  astyle code format
  write localization guide with msvc
  use unicode escape character to represent single char '…'
  fix min(pow(x,x),x) build error with gcc
  astyle explosion.cpp and output.cpp/h
  add comment and fix the typo
  add SDL into msvc compiling guide
  write compiling guide for visual studio
  solve almost all compile errors with msvc
  fix constant integer array issue with MSVC in input.cpp
  add msvc140 cache files into gitignore

+++ src/bionics_ui.cpp
@@ -73,7 +73,7 @@ void draw_bionics_titlebar( WINDOW *window, player *p, bionic_menu_mode mode )
 
 const auto separator = []( std::ostringstream &s )
 {
-    return s.tellp() != 0 ? ", " : "";
+    return ( int )( s.tellp() ) != 0 ? ", " : "";
 };
 
 //builds the power usage string of a given bionic
